% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_allowed_values.R
\name{check_allowed_values}
\alias{check_allowed_values}
\title{Check Allowed Values}
\usage{
check_allowed_values(data, variable, allowed_values)
}
\arguments{
\item{data}{A data frame.}

\item{variable}{Character name of the variable to check.}

\item{allowed_values}{Vector of permitted values.}
}
\value{
A data frame containing records with invalid values.
}
\description{
Identifies records containing values that are not included
in a specified set of allowed values.
}
