% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicates.R
\name{check_duplicates}
\alias{check_duplicates}
\title{Check Duplicate Records}
\usage{
check_duplicates(data, variable)
}
\arguments{
\item{data}{A data frame.}

\item{variable}{Character name of the variable to check.}
}
\value{
A data frame containing duplicated records.
}
\description{
Identifies duplicated values for a specified variable.
}
