% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing.R
\name{check_missing}
\alias{check_missing}
\title{Count Missing Values}
\usage{
check_missing(data, variable)
}
\arguments{
\item{data}{A data frame.}

\item{variable}{Character name of the variable to check.}
}
\value{
Number of missing values.
}
\description{
Counts the number of missing values in a specified variable.
}
